fix: separate creation of browser notifications and sounds from SFC component
